Removed certain threading code from our main window.
Although having a list model and its widget in different threads worked
"okay" in Qt4, in Qt5 this doesn't work at all. As a result, we need to
alter our application to use QtConcurrent's QFuture API instead.

This commit removes the old broken threading code, so that our
application will "work" (albeit by locking the GUI when performing long
tasks), so that the new threading code can be added on.